Initial commit
This sets up a lab environment ready for AZs, which is aimed to
contain 3 AZs, all sharing the same network L2. This allows the
current deployment model without large changes.

This is a good way to exercise the first step of AZ conversion,
by modifying a bit of our configuration.