Clarify usage of tfvars
Instead of adding my key into default, it makes more sense to
fail when a key is not set, and ask the user to define its own
keys.

For that, we should use the tfvars auto load feature.
I have clarified, with a README, how this whole terraform can be
used.